Add alt code to 2014/birken + format/typo checks
The alt code lets one redefine the port and also redefine the timing
constants. The reason for the port is in case of firewall (though on
localhost that shouldn't be a problem but it does appear to be that it
might bind to other addresses too). The timing one the author specifies
some uses for changing it. Uses #ifndef..#endif.

Format/typo check README.md file.