Add alt code to 2014/maffiodo2 + try.sh script
The alt code is from the author.

The try.sh script has the example command provided by the judges as well
as some by the author. One was removed because of a syntax error
(removed from the author's examples).

Format/typo check README.md file.

The only thing that has to be done is to update the references to
twitter to note how the IOCCC no longer endorses twitter or whatever the
rocket destroyer wishes to call it (though certainly it won't be worded
like that :-) ).