Add 2015/mills1/try.sh + format/typo checks
The try.sh script runs the make command in the judges' remarks to make
it easier.

A warning is that if you have good eye-hand coordination and you're good
at games and you have too much time you might end up falling asleep
playing it as you'll probably never die. I only died because I was
trying to mark a reminder complete on the other side of the screen at
the same time as playing .. and I almost survived that .. and 'Cody's
also blind' as the joke goes in the family. That's probably why I ended
up dying. At that point I already had over 90 points.

Still the purpose is to make it easier to play if the default settings
are too frustrating. I contemplated playing with other settings but I
decided not to. An example where it might be improved is to let the user
pass in the environmental variables to override should they wish to but
it is not even clear to me how the variables should be sanitised so
that's an argument against it.