Improve redefining sleep time in 2000/thadgavin
Now one can do instead

    make SLEEP=555 SDL_SLEEP=555 alt

where SLEEP is the variable for the curses version and SDL_SLEEP is for
the SDL version. One need not specify both if they do not want to
redefine one or the other and of course one doesn't have to specify a
sleep time at all either.