Add 2015/dogon alt code + other improvements
The author suggested that one should change the value of 'q' to
'0x82000820000000' in order to see a bug that they carefully avoided. To
make it easy to do this the alt version was added.

The Makefile has been improved so it's easier to change the dimensions
of the game: the author had put in the Makefile two macros that have the
width and height and these now are set to WIDTH and HEIGHT defined in
the Makefile so that one can just do something like:

    make WIDTH=500 HEIGHT=500 all

rather than being forced to update the Makefile to do this.

The README.md file has been format/typo checked.