Add alt code to 2020/ferguson2
This alt code prints out the text one character at a time to make it
easier to read and more generally so it's not dumped out at once. This
also makes it a bit more like the original thing (at least as I recall
it :-) ).

The default sleep time is 400000 microseconds but one can change it
like:

    make clobber SLEEP=0 alt

which would disable the delay, making it like the original.

If it's < 0 then it's set to the default.

I did not yet add a try.sh script. That'll likely come later when
working on 2020. I only did this now as I thought of this idea and
didn't want to forget it.