Fix assumption that EOF==-1 in 2015/schweikhardt
The author stated EOF must be -1 but this is, as the author even noted,
not guaranteed by the standard (it just has to be a value < 0).

But rather than add C pre-processor directives to force EOF to be -1
(preferred over changing EOF to -1 in the code) the Makefile does (after
${CFLAGS}) -UEOF -DEOF=-1 so as to not have to modify the code any.

A few fixes in the README.md file.

This should complete this entry, leaving just (I think) 2015/endoh1.