Add try scripts to 2018/ferguson + formatting
The try.sh and try.alt.sh scripts are respectively for the entry and the
alt code. There is a joke in them that is based on what the judges
originally suggested one try.

The weasel.man page was deleted. Instead weasel.1 has had the FILES
section removed (which caused some confusion in another way with
another winner of that year). The spec file will have to be tested again
another time.

I did NOT format / typo check the README.md file. Too much to do for now
and since it's mine I can do it any other time. I'd like to get through
the rest of the entries.

Thus this should complete (for now?) 2018/ferguson.