improve the consistency of package install related FAQ
We make how "how to install" package XYZZY more consistent
for those FAQ in terms of the order the OS/methods
as well as mentioning all of the current methods:

* Red Hat based Linux
* macOS
* mavOS via Homwbrew
* mavOS via macports
* Debian based Linux
* Other Linux distributions
* package website

The sound based FAQ was modified in a similar way
so that should specific Linux distribution instructions
need to be added, those could be added under the
appropriate section.

Use only "dnf install" for RHEL and mention the possibility of that
older RHEL distributions might have to use "yum" instead of "dnf"

Changed linux to Linux.

Suggest in FAQ 3.1 that they look at other FAQ entries (in section 3).

Added missing empty lines before markdown headers.