Add 2019/burton/try.sh and other fixes/updates
The try.sh runs prog and wc side by side but in a way that is formatted
(including alignment) so it's easier to see what line is from which
program and it also runs make test.

The script cow was renamed cow.sh to help browsers with whether or not
to display or download.

The script check was renamed check.sh to help browsers with whether or
not to display or download.

The script check.sh was fixed in case the 'x' (passed in to the script
as an environmental variable) is not executable (it resets it back to
./prog) and also made to satisfy ShellCheck.

The script cow.sh was fixed for ShellCheck moo (or was it too? :-) ).