Add FAQ about dead and invalid links
... and what one should do, of course. This includes an example of an
invalid link (albeit an extreme one :-) ) that I fixed quite some time
back (months ago) just to give an idea. But it's more useful to check
for dead links in the IOCCC website itself (linking to pages/files in
the website that is). Of course I'll be doing that once the README.md
files are in order but there's a high chance I won't be able to find
them all esp as after the website is done I have more to do.