add instructions for how to compile 1988/spinellis
Because of the nature of 1988/spinellis, we cannot compile
the entry directory.  While there may be workarounds to
automatically compile code that tries to read source from
standard input, such workarounds are likely not portable
or at best uglify the process.

NOTE: This entry should NOT use a `try.sh` nor `try.alt.sh`
approach.  Instead the user needs to be directed to perform
the actions.  We do this by echoing instructions from the Makefile.

We remove the entry line from `input.txt` and `input2.txt`
to make the echoed instructions in improve the clarity
of the echoed instructions.

We turn off all warnings, due to the fact that whatever
the user might enter might have minor issues.  As such,
and to be similar, we turn off all warnings to the compiled
commands just in case compiler writers or ANSI C does
something silly in the future.  ;-)

We clean up `make clean`, `make clobber` as well as
updating and sorting `.gitignore`.