Fixes in 1987/lievaart
The optimiser was left disabled by accident.

Check for !isdigit(*A) before using atoi(3) on A. This was part of the
bug fixes I made (infinite loop flooding the user with invalid input
which is frequent if one does not know how to play or if one makes a
typo). This is in both versions, the one with the board and the entry
itself (the board being too big to fit in the contest size limits of the
time).