Fix 1990/jaw/Makefile, improve try.sh + bugs.md
The optimiser was left disabled by accident.

The try.sh now adds another file in the archive.

Add entry to bugs.md. There are two issues, or at least one. One issue
might not be an issue but it appears to be. The one that appears to be
was supposed to be added to the bugs.md file ages ago but was missed by
accident.