Add 2020/carlini/try.sh, update bugs.md
Why a try.sh for this? It shows that you can also automate the play (it
happens to come from the author's remarks, however, so there's that
too). Then it tells you to be serious by playing with a friend, either
real or imagined, rather than doing something so silly.

Update bugs.md with a INABIAF section.

The README.md file has been format/typo checked.