Add try scripts to 2020/ferguson2, other updates
The try.sh script is for the original entry. The try.alt.sh is of course
for the alt version that I only recently made.

Removed html files recode.html and chocolate-cake.html (the horror! But
there is the markdown file so all is good... or would be if it wasn't
for the fact it's enciphered as a fun way to experience the entry .. or
else otherwise a frustration if there wasn't
2020/ferguson1/chocolate-cake.md? :-) ).

Updated .gitignore.

Minor updates to chocolate cake file: but of course for most people they
might want 2020/ferguson1/chocolate-cake.md instead.

This should be all that needs to be done for 2020/ferguson2 - for now
anyway. The documentation has not been checked for typos/formatting but
that might happen another time.