Add 2020/kurdyukov1/try.sh + format/typo checks
The try.sh script tries to find md5sum and if it does not succeed it
tries for openssl. If it is openssl it uses the subcommand md5;
otherwise it just uses md5sum. Whatever the case if either is found it
writes the checksum of prog.c to md5.txt and then it uses prog to write
the checksum to prog.md5.txt. Then it uses diff -s on the two. If
neither tool is found it just skips that check.

Format/typo check README.md file.