Fix 2004/gavin/img/fs.tar
Now contains old hint file (which is a copy of README.md) and the
updated README.md file. This makes it match correctly. This has not yet
been removed from the todo file as other things are also being removed
from the todo file and I'd rather do those changes together as a single
commit of things completed.

Minor update to thanks file wrt this entry.