Add try.sh to 1985/august
The script has fun with reversing bits too through rev(1) (if
installed).

The script calls primes.sh even if primes(6) is not installed as that
script will check that it's installed and will tell one where to get it
if it's not. It only uses the default value, however, so primes.sh is
still shown in the README.md file.