Remove item about usleep macro name from todo file
In alt code where a call to usleep(3) is made to allow for configuring
speed originally I used the macro 'Z'. In one case 'Z' was already taken
so I changed it to 'S' for that entry. I had the thought it might be
nice to make it always the same. This however is not possible as more
entries already have 'S' used. Also it's not important since I changed
the Makefiles so that one can specify it like 'make clobber SLEEP=500
alt' or whatever, removing the macro name and having to mess with
'CDEFINE'.