Remove 1986/wall alt code
The alt code was just a patched (by Yusuke Endoh) version (two strings
were strdup()d) that still required -traditional-cpp. He is still
credited in the thanks file but one can do make diff_orig_prog to see
what was changed to make it not require -traditional-cpp and Yusuke's
fix is there too. The thanks file details what I had to do to make it
work without -traditional-cpp and his change is still there as well. It
seemed to serve no purpose to have an alt version that requires one to
use -traditional-cpp when one can just look at the original code if they
want to see what it was like back in 1986.