Typo fix in thanks file
When moving honour roll to the end of the file some things that refer to
'below' have to be changed to 'above'. It's possible more need to be
found but I've not looked at it yet. I haven't actually gone through the
file (that can come later); I just had the file open in GitHub and my
eyes zoomed in on the error.