improve readme2index.sh option checking, explain restrictions
The `readme2index.sh` tool does a more consistent job of checking
options in the help message and reporting when something violates
a restriction.

In many cases the parameters (_$OPTARG_) of an option is not allowed
to contain a single-quote, nor a double-quote.  We prevent use of
single-quotes and double-quotes in cases where passing such to other
tools may create a problem.

There are a few exceptions to the above mention "single-quote /
double-quote" restrictions.  For example, the "-s token=value" does
NOT have such a restriction.  In the case of  "-s token=value", we
can allow for the single-quotes and double-quotes because the ;
(semicolon) is used as a regex separator instead.

We also better explain the cumulative nature of "-s token=value".

Some help message wording cleanup was performed.

The `inc/md2html.cfg` file was updated so that the comment
about options matches the `readme2index.sh` tool usage message.