Add 1987/hines/try.sh
The files goto.c and goto.txt were also added. The first one is a short
file that is easy to see how many 'goto's there are and the goto.txt
file has seven 'goto's and one 'Goto' to demonstrate that the program is
case-sensitive.

The .gitignore file was updated with 'goto' though the Makefile does not
have (of course) a 'goto' rule as it's only there to have a C file which
is easy to see how many 'goto's there are.