Fix how 1993/lmfjyh alt code is built
In the case that the original entry fails to compile (if you don't have
gcc < 2.3.3 this will happen) the alt code will automatically be built.
One can still manually compile the alt code, of course, but this way one
can at least use the entry. Note that in the case the alt code is
automatically compiled it is compiled to the lmfjyh just like what I did
with 1985/sicherman. This was discussed after I did the 1985/sicherman
(which will compile anyway after my fixes unless a compiler objects to
main() having only one arg).