Make corrections with 2006/toledo2 wrt files
There actually is a file missing. It is not in the archive: IMPORT.COM.
Thus it has been added back to bugs.md until such a time as we can find
it or be provided it.

The .COM files should not be links. This has been discussed and it was
agreed to not link to them as they'll just be downloaded and they
already are in the repo.

Removed the zip file as the files have already been extracted. This is
how it was done with other entries too and this is what was desired as
well.