Rename 1994/westley/spoiler.sh try.sh
The script now first checks columns (tput cols) and if it's not 80 it
will ask the user if they still wish to continue. Checking the tab
length of 8 and wraparound is probably not worth the effort. Also
unfortunately output is not as nice as it used to be because compilers
show more context with errors. Still having columns at 80 does improve
the output a fair bit.