Improve alt build of 1989/ovdluhe
One can now do

    make clobber P=5 alt

instead of the more awkward

    make clobber CDEFINE="-DP=5" alt

The try.alt.sh script was added and allows one to do

    P=5 ./try.alt.sh

If P is not specified it uses the default of the alt code (which checks
if it's defined first via #ifdef..#endif). Otherwise it uses the new
value. Of course if one specifies not an int or too high value (or maybe
too low - didn't try that) the program will fail. That is the fault of
the user. The author does note a reasonable range to try.