Add 1991/brnstnd/try.sh and try.txt
The try.txt is a file for the rot13 SORTA file. All but one of the
sorta.* files is used. The only one that is not prints out a verr long
line of output and would flood the screen. The try section in the
README.md does suggest one try it just to be complete.