Add 1994/horton/try scripts and article cited
The try.sh and try.alt.sh scripts correspond to the original entry and
the alt code (which is enhanced).

The PDF files might need a bit more explaining. The author wrote an
article in 'The USENIX Association Newsletter ;login: September /
October 1992 Vol. 17, No. 5' (the exact title from the PDF) called
'Using GraphTab'. I found the full newsletter in the Internet Wayback
Machine which is now added to the repo (okay: the newsletter as a PDF is
added to the repo, not the Internet Wayback Machine :-) ). But the pages
cited are 28-31 so I extracted (with macOS Preview app: File -> Print
option) the specific pages so one need not look at the full newsletter
if they do not wish to. It felt better to have the PDF rather than
images as it's easier to deal with what the right size is: the size is
there and one can zoom in or out as they see fit (whereas with images
different resolutions make it easier or harder to see what is in them).