Add 1994/imc/try.sh + inclusion of unistd.h
The inclusion of unistd.h is for getpid(2). While not strictly necessary
(at least not observed) this type of thing can be a problem and in some
other entries a missing or incorrect prototype caused the entry to fail
(sometimes only some systems namely macOS). This was just noticed by
chance and I did it on a whim.