Remove reference to jargon file in 2000/rince
The bugs.md file is all that's needed; the information is already there
in bugs.md and the purpose of having the 'STATUS: ...' lines in the bugs
and (mis)features section of the README.md files was to not have that.
Still having the reference to DISPLAY being unset being a problem seems
like a good idea.