Remove grawlixes from 2012/vik/README.md
We all know what is meant by 'brain$#@$' and there's no need to censor
it especially when it links to the article on wiki that does not censor
it (not that grawlixes are even censoring anything). This means there's
no need to escape the '$' as they are not there now.