Fix references in 2018/burton1/README.md
I believe that the previous commit got the ordering wrong. The
significant inspiration for the contest was linked to directly in
markdown rather than a '[...]: ...' style link so reference 3 should
actually be reference 2 (instead of the link to the main IOCCC readme
file). A question of course is what to do about the link

    http://ioccc.org/all/README

since all README files have been changed to README.md files.
Nevertheless right now the ioccc.org site has README, not README.md, so
it would be either a broken link or it would refer to say /all/README.md
(which might be the best way?).