Make my fix to 1995/makarios more like original
The parameters to the functions spell out (backwards) the function name.
As for main() since there can only be three args it has in comments a
fourth to spell out main (backwards). Also use old style
( main(n,i,a)char**i,**a;{/*..*/} ) as that's how the original code is
(and the added function is that way too though only ints like main()
used to be).

Added -Wno-implicit-int -Wno-return-type to CSILENCE in Makefile.
There's still one more warning but I'm not sure if it's supported by all
compilers so I haven't disabled it (yet?).