Fix comment about 1996/schweikh2/try.sh
The script was changed to run a certain command only once but the
README.md file stated the old way of it running several times in a row.
I can't see the use of running it several times when we can just pipe it
through less to see the look. The author's suggested command, meanwhile,
is hard to see what is happening as it doesn't go slow enough and it
floods the screen until one hits ctrl-c/sends intr.