Add 1998/df/try.sh
This script runs repeatedly until (when prompted) one hits 'q' (or 'Q')
or else sends ctrl-c/intr. I also added a question to the extra credit
questions: are there any misspelt words and if there are what are they?
I also added that there's a way to cheat very easily but I did not say
how to do it. It should be pretty obvious how to do it to anyone who
tries the program once or if not once a few times.