Add 1998/bas1/bas1.sh
This simplifies the running of the program as the command lines
(different options) can be long and more complicated to type, especially
the last one. The script is menu driven. Initially I made it a loop so
one can run the command once to try multiple modes but this seems
unnecessary and created some display problems (depending on how
invocations were done).