Fix typo (bad filename) in 2004/arachnid
I had renamed the arachnid.info to arachnid.txt as it's not really an
informative file but I forgot to update the command in the README.md
file. I was tempted to rename it to arachnid.maz as it's a maze file but
I kept it as a text file so that (some?) browsers know what to do with
it (though I'm not sure how strictly necessary that is it doesn't hurt
to have it as .txt either).