Silence 1992/gson/mkdict.sh shellcheck warnings
As a test of the new format of disabling warnings/errors from
shellcheck(1) I have done it for this script which is an important one
to consider as every warning and error should be disabled as the author
provided the script (that I added as a script file) as an obfuscated
script and therefore to not follow best practices, not at the time or
now. But this way it's just one script to see if this works well, rather
than fixing all the others (which will happen later in any case).