Add 1994/westley alt code (for easier use)
As 1994/westley requires (for good display) 80 columns, 8 character tabs
and wraparound it makes it harder to run the entry to enjoy it with
modern systems. Also because it generates messages through error
messages (of the compiler line) different compilers will show it
differently and it can end up being harder to follow. The alt code uses
the same logic but instead of using #includes (of files that don't
exist) it uses printf(3). Everything else is the same.

The try.alt.sh is just like the try.sh but for the alt code.

Since the alt code is ideal for modern systems the original code is
suggested second. This can be changed of course.