Further improvements to 1996/gandalf/try.sh
Running ./try.sh by itself will not show this improvement. It might be
an inside joke between Landon and me but which adds fun to the entry in
a way that many people will appreciate, given the entry name. Those who
do not, however, will not be bothered by it as running ./try.sh won't do
anything different from before: it'll just run the program in a variety
of ways as before.