Slight UI improvements in 1996/gandalf/try.sh
Rather than showing something like:

    Showing diff between 'cmd' and 'cmd':
    Press any key to run: diff ...

it now shows instead (for example):

    $ diff hatcat.txt cathat.txt (diff between 'hatcat' and './cathat'):
    Press any key to continue:

The reason it's two lines is because having it on one line makes it a
kind long line. Probably not an issue with modern systems but I have
tried to keep the lines close to 80 columns and not doing it this way
would make it almost 90.

The script now has -C 1 used to show context. This way the file names
are shown in the output to help distinguish the files being compared. Of
course it might be possible to diff the command outputs themselves but
this works better because it will show the files that are the same
(when they are) rather than '.. and - are identical'.