Improve 2012/grothe/try.sh + other minor fixes
The try section in the README.md file did not refer to the script.

The URL in the cookie recipe file was updated to refer to the domain
that the old one redirects to just in case the old one ever expires or
no longer redirects to the more recent domain. The thanks file notes
this for reference as an aside for context and historical purposes,
noting both URLs.