Improve 2012/konno/try.sh
Now randomly selects 10 numbers (in range $((RANDOM % 101))).

Also allows one to pass in numbers (or if they wish to they can pass in
non-numbers) to pass to the program.

The script has new fun numbers.

Removed that the valid range is from 0 - 255 as numbers bigger than 255
not only work but some show some fun output (a number of which are used
in the script now).