Improve scripts of 2013/dlowe
The try.sh script cannot be modified to prompt the user to run commands
as the other scripts (or at least one, diff.sh) make use of tail(1) on
the output. The script diff.sh however can have that and it has been
done.

Minor fixes in display in slflen.sh.